requirements.txt
fastapi[standard]
uvicorn
paho-mqtt
python-dotenv
sqlalchemy
psycopg2-binary
sqlmodel
pydantic